Move Composable View to separate package
