+++ app/build.gradle.kts
@@ -62,12 +62,15 @@ dependencies {
     implementation("androidx.core:core-ktx:1.13.1")
     implementation("com.google.android.gms:play-services-mlkit-face-detection:17.1.0")
     implementation("com.google.mlkit:camera:16.0.0-beta3")
+    implementation("androidx.activity:activity-ktx:1.9.2")
     testImplementation("org.mockito.kotlin:mockito-kotlin:5.4.0")
     testImplementation("org.robolectric:robolectric:4.12.1")
     testImplementation("junit:junit:$jUnit4Version")
     testRuntimeOnly("org.junit.vintage:junit-vintage-engine:5.11.0")
     androidTestImplementation("androidx.compose.ui:ui-test-android") // Compose test framework
     androidTestImplementation("androidx.test.espresso:espresso-core:3.6.1")
+    androidTestImplementation("androidx.compose.ui:ui-test-junit4:1.7.1")
+    androidTestImplementation("androidx.test.ext:junit:1.2.1")
     androidTestImplementation(platform("androidx.compose:compose-bom:2024.06.00"))
     debugImplementation("androidx.compose.ui:ui-tooling")
     debugImplementation("androidx.compose.ui:ui-test-manifest") // Needed for createComposeExtension()
