+++ app/src/main/java/com/example/safedistance/utils/ServiceCommands.kt
@@ -1,10 +1,10 @@
 package com.example.safedistance.utils
 
-import android.content.ContextWrapper
+import android.content.Context
 import android.content.Intent
 
 class ServiceCommands(
-    private val context: ContextWrapper
+    private val context: Context
 ) {
     fun sendServiceCommand(action: String, serviceClass: Class<*>) {
         val intent = Intent(context, serviceClass).apply {
